Adding the one python file.
This file reads through a given user's gmail account for emails received by a particular sender. Based on the body of those emails, the program can decide whether or not the email contains a 'job' that needs to be created. It then creates an event on the user's Google Calendar using API and a Google Sheet (using Google Sheet API and gspread) to help organize the data.